FocusNest v1.0 BETA
The website is finally here. The Home Page is ready with everything and is now out i am now working on the settings page which should be done in no time aswell. And some bug fixes were there and also now when a task finishes a sound is there. will plan to add notifications aswell.